♻️ 2V5SZJ code: split blueprint hotspot modules
