+++ packages/agentplane/src/runner/usecases/task-run.ts
@@ -38,6 +38,8 @@ import { persistRunnerOutcomeToTask } from "../task-state.js";
 import { RunnerRunRepository } from "../run-repository.js";
 import { createRunnerRunId, resolveTaskRunnerPaths } from "../task-run-paths.js";
 import { normalizeRecipeArtifactPrefixes } from "../result-manifest-policy.js";
+import { renderTaskRunnerBootstrap } from "./task-run-bootstrap.js";
+export { renderTaskRunnerBootstrap } from "./task-run-bootstrap.js";
 import {
   RUNNER_API_VERSION,
   RUNNER_BUNDLE_SCHEMA_VERSION,
@@ -151,7 +153,7 @@ export function assertRunnerBlueprintPolicyModuleBudget(bundle: RunnerContextBun
   const actualCount = Math.max(policyModules.length, policyManifestEntries.length);
   if (actualCount <= maxPolicyModules) return;
   throw new CliError({
-    exitCode: 2,
+    exitCode: exitCodeForError("E_VALIDATION"),
     code: "E_VALIDATION",
     message: [
       "Runner blueprint policy module budget exceeded.",
@@ -220,7 +222,7 @@ function resolveRunnerBlueprintPlan(opts: {
     })
     .catch((err) => {
       throw new CliError({
-        exitCode: 2,
+        exitCode: exitCodeForError("E_VALIDATION"),
         code: "E_VALIDATION",
         message:
           err instanceof Error
@@ -353,52 +355,6 @@ export function assertRunnerTaskExecutable(bundle: RunnerContextBundle): void {
       `${task.task_id}: runner execution requires task status DOING ` +
       `(current=${JSON.stringify(status)}; use \`agentplane task start-ready ${task.task_id} --author <ROLE> --body "Start: ..."\` first).`,
   });
-}
-
-export function renderTaskRunnerBootstrap(
-  bundle: RunnerContextBundle,
-  invocation?: RunnerInvocation,
-): string {
-  const targetLabel =
-    bundle.target.kind === "task"
-      ? `task ${bundle.target.task_id}`
-      : `recipe scenario ${bundle.target.recipe_id}:${bundle.target.scenario_id}`;
-  const stopRules = bundle.blueprint?.stopReasons ?? [];
-  return [
-    "# agentplane runner bootstrap",
-    "",
-    "This invocation is already inside an approved runner execution.",
-    "- Do not run repository startup commands such as `agentplane config show`, `agentplane quickstart`, `agentplane task list`, `git status`, or `git rev-parse` unless the bundle explicitly requires them as task work.",
-    "- Do not create, approve, start, verify, finish, block, or rerun tasks unless the bundle explicitly requires task metadata edits.",
-    "- Do not recursively invoke runner entrypoints such as `agentplane task run` or `agentplane recipes scenario execute` from inside this run.",
-    "- Open bundle.json immediately, execute the requested work directly, and stop when the requested outcome is satisfied.",
-    "",
-    `- target: ${targetLabel}`,
-    `- adapter: ${bundle.execution.adapter_id}`,
-    `- mode: ${bundle.execution.mode}`,
-    `- run_id: ${bundle.execution.run_id}`,
-    `- bundle_path: ${bundle.execution.artifact_paths.bundle_path}`,
-    `- result_path: ${bundle.execution.artifact_paths.result_path}`,
-    `- bootstrap_path: ${bundle.execution.artifact_paths.bootstrap_path}`,
-    "",
-    "Use bundle.json as the complete runner input. Do not reconstruct prompts from CLI argv.",
-    ...(stopRules.length > 0
-      ? [
-          "",
-          "Blueprint stop rules:",
-          ...stopRules.map((rule) => `- ${rule.severity}: ${rule.reason} (${rule.id})`),
-        ]
-      : []),
-    "Execute-mode runs must write a valid JSON result manifest to result_path before exiting.",
-    "Minimal manifest example:",
-    '{"schema_version":1,"status":"success","summary":"Completed.","capabilities_used":["runner.exec"]}',
-    "",
-    "Prepared invocation:",
-    "",
-    invocation
-      ? `- argv: ${invocation.argv.join(" ")}`
-      : "- argv: <not prepared; preflight refused>",
-  ].join("\n");
 }
 
 export async function prepareTaskRunnerExecution(opts: {
