+++ scripts/oversized-test-baseline.json
@@ -3,7 +3,7 @@
   "threshold_lines": 1000,
   "budgets": {
     "max_entries": 10,
-    "max_total_lines": 11311,
+    "max_total_lines": 11429,
     "max_file_lines": 1300
   },
   "entries": [
@@ -29,7 +29,7 @@
     },
     {
       "file": "packages/agentplane/src/cli/run-cli.core.tasks.query-listing.test.ts",
-      "lines": 1183
+      "lines": 1190
     },
     {
       "file": "packages/agentplane/src/cli/run-cli.core.test.ts",
@@ -45,7 +45,7 @@
     },
     {
       "file": "packages/agentplane/src/commands/task/finish.validation.unit.test.ts",
-      "lines": 1103
+      "lines": 1218
     }
   ]
 }
